requirements.txt
pillow
setuptools
-e ./tools/
